requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
boto3==1.9.251
botocore==1.12.251
chardet==3.0.4
docutils==0.15.2
idna==2.8
jmespath==0.9.4
multidict==4.5.2
python-dateutil==2.8.0
s3transfer==0.2.1
six==1.12.0
slackclient==2.2.1
urllib3==1.25.6
yarl==1.3.0
